Fixed Scenario module, was causing nullrefs and preventing anything from being done in the game
