+++ Changelog.txt
@@ -1,5 +1,8 @@
 ChangeLog
 
+0.5.3.2
+	Fixed Scenario module, was causing nullrefs and preventing anything from being done in the game
+
 0.5.3.1
 	Fixed initialization of vars
