+++ LaunchNumbering.version
@@ -10,7 +10,7 @@
     "MAJOR": 0,
     "MINOR": 5,
     "PATCH": 3,
-    "BUILD": 1
+    "BUILD": 2
   },
   "KSP_VERSION": {
     "MAJOR": 1,
